+++ shuup/xtheme/plugins/text.py
@@ -5,12 +5,10 @@
 #
 # This source code is licensed under the AGPLv3 license found in the
 # LICENSE file in the root directory of this source tree.
-import markdown
-from django import forms
-from django.utils.html import escape
 from django.utils.safestring import mark_safe
 from django.utils.translation import ugettext_lazy as _
 
+from shuup.admin.forms.widgets import TextEditorWidget
 from shuup.xtheme.plugins._base import Plugin
 from shuup.xtheme.plugins.forms import TranslatableField
 
@@ -25,15 +23,10 @@ class TextPlugin(Plugin):
         ("text", TranslatableField(
             label=_("text"),
             required=False,
-            widget=forms.Textarea,
-            attrs={"class": "remarkable-field"}
+            widget=TextEditorWidget
         ))
     ]
 
     def render(self, context):  # doccov: ignore
         text = self.get_translated_value("text")
-        try:
-            markup = markdown.markdown(text)
-        except:  # Markdown parsing error? Well, just escape then...
-            markup = escape(text)
-        return mark_safe(markup)
+        return mark_safe(text)